[refactor] 스토리북 파일 추가 #113
